requirements.txt
coingecko-sdk
python-dotenv
psycopg2-binary
apache-airflow-providers-postgres